mock_encointr_api.dart line 108

+++ lib/mocks/substrate_api/mock_encointer_api.dart
@@ -105,7 +105,7 @@ class MockEncointerApi extends EncointerApi {
 
   @override
   Future<DateTime?> getMeetupTime() async {
-    return DateTime.fromMillisecondsSinceEpoch(claim['timestamp']);
+    return DateTime.fromMillisecondsSinceEpoch(claim['timestamp'] as int);
   }
 
   @override
